lib: bugzilla: use actual get request instead of search endpoint for `get`
Otherwise, using the search endpoint hides various types of bugs from
view that are propagated by the `get` endpoint. For example, nonexistent
bug IDs errors are raised while they are ignored entirely for the search
endpoint.

The new method still supports multiple bugs in a single request, it's
just an undocumented way of doing that for the REST API. This is
probably due to it existing for the XML-RPC and/or JSON-RPC interfaces
and some of the codepaths remain the same for the newer interface.